move the handler one component down
