+++ web/playground/src/components/editor/Editor.js
@@ -1,4 +1,4 @@
-import { useState, forwardRef } from "react";
+import { useEffect, useState, useRef } from "react";
 import { default as MonacoEditor } from "@monaco-editor/react";
 import "./Editor.scss";
 
@@ -10,9 +10,27 @@ import "./Editor.scss";
  * program written in the MonacoEditor
  * @returns
  */
-const Editor = forwardRef(function Editor({ onEditorChanged, initialValue }, ref) {
+function Editor({ onEditorChanged, initialValue }) {
   const MS_TIMEOUT_TO_EVALUATE = 5000;
   const [lastModifiedTimeout, setLastModifiedTimeout] = useState(() => {});
+  const [program, setProgram] = useState(initialValue);
+  const editorRef = useRef(undefined);
+
+  const handleKeydownEvent = (event) => {
+    event.stopImmediatePropagation();
+    if (event.key === "S" || event.key === "s" && (event.ctrlKey || event.metaKey)) {
+      // handle save action
+      onEditorChanged(program);
+      event.preventDefault();
+    }
+  };
+
+  useEffect(() => {
+    editorRef?.current?.addEventListener("keydown", handleKeydownEvent);
+    return () => {
+      editorRef?.current?.removeEventListener("keydown", handleKeydownEvent);
+    };
+  });
 
   /**
    * Sets a timeout that will pass the latest modification to the parent component
@@ -20,23 +38,24 @@ const Editor = forwardRef(function Editor({ onEditorChanged, initialValue }, ref
    * or more modifications are made between 0..MS_TIMEOUT_TO_EVALUATE then it clears
    * the timeout and sets another one with the updated value
    *
-   * @param {string} program
+   * @param {string} programInEditor
    */
-  const handleEditorChange = (program) => {
+  const handleEditorChange = (programInEditor) => {
+    setProgram(programInEditor);
+
     if (lastModifiedTimeout) {
       clearTimeout(lastModifiedTimeout);
     }
 
     setLastModifiedTimeout(
       setTimeout(() => {
-        console.log(program);
         console.log(`${MS_TIMEOUT_TO_EVALUATE} without changes, sending the back to parent component`);
         onEditorChanged(program);
       }, MS_TIMEOUT_TO_EVALUATE));
   };
 
   return (
-    <div className="Editor" ref={ref}>
+    <div className="Editor" ref={editorRef}>
       <MonacoEditor
         height="100vh"
         defaultLanguage="javascript"
@@ -51,6 +70,6 @@ const Editor = forwardRef(function Editor({ onEditorChanged, initialValue }, ref
       />
     </div>
   );
-});
+}
 
 export default Editor;
