+++ web/playground/src/screens/playground/Playground.js
@@ -1,6 +1,6 @@
 import Editor from "../../components/editor/Editor";
 import Shell from "../../components/shell/Shell";
-import { useState, useCallback, useMemo, useEffect, useRef } from "react";
+import { useState, useCallback, useMemo, useEffect } from "react";
 import { useParams, useLocation } from "react-router-dom";
 import { getPlaygroundHistory } from "../../services/PlaygroundService";
 import useWebSocket from "react-use-websocket";
@@ -12,8 +12,6 @@ function Playground() {
   const [program, setProgram] = useState("");
   const [history, setHistory] = useState([]);
   const [playgroundNotFound, setPlaygroundNotFound] = useState(false);
-
-  const editorRef = useRef(undefined);
 
   const { sendMessage, lastMessage } = useWebSocket(WEBSOCKET_URL);
 
@@ -28,17 +26,6 @@ function Playground() {
       setHistory((prev) => prev.concat(evalData.result));
     }
   }, [lastMessage]);
-
-  const handleKeydownEvent = (event) => {
-    event.stopImmediatePropagation();
-    if (event.key === "S" || event.key === "s" && (event.ctrlKey || event.metaKey)) {
-      // handle redo action
-      console.log("saved");
-      event.preventDefault();
-    } else if (event.key === "Z" && (event.ctrlKey || event.metaKey)) {
-      // handle undo action
-    }
-  };
 
   useEffect(() => {
     (async () => {
@@ -56,8 +43,6 @@ function Playground() {
           // of a known value
           handleEditorChanged(program);
         }
-
-        editorRef?.current?.addEventListener("keydown", handleKeydownEvent);
       } catch (error) {
         setPlaygroundNotFound(true);
         return;
@@ -66,7 +51,6 @@ function Playground() {
     return () => {
       // make sure location state is cleared on browser refresh
       window.history.replaceState({}, document.title);
-      editorRef?.current?.removeEventListener("keydown", handleKeydownEvent);
     };
   }, [state?.skipFetchHistory]);
 
@@ -96,7 +80,6 @@ function Playground() {
         <Editor
           onEditorChanged={handleEditorChanged}
           initialValue={program}
-          ref={editorRef}
         />
       </div>
       <div style={{width: "30vw", height: "100vh"}}>
